requirements.txt
requests==2.31.0
python-docx==1.1.2
yfinance==0.2.28
vnstock==0.1.1
ta==0.10.2
streamlit==1.36.0
pandas==1.3.5
plotly==5.16.0
click==8.0.4
prophet==1.1.5
numpy==1.23.4
scikit-learn==1.5.1
boto3==1.34.143
botocore==1.34.143
bs4==0.0.2
langchain==0.2.7
langchain-community==0.2.7
python-dotenv==1.0.1
langchain-aws==0.1.10
streamlit-chat==0.1.1
PyPDF2==3.0.1
pandas-datareader==0.10.0
anthropic==0.31.0
vnstock3==0.3.0.6
seaborn==0.13.2
transformers==4.42.3
